+++ NEWS.md
@@ -1,4 +1,6 @@
-# Version 1.1.0.9000
+# Version 1.1.0.9001
+
+* Memoise now uses caching objects from the cachem package by default. These caches support automatic pruning, so that they won't grow indefinitely. The older-style cache objects in the memoise package are still supported, but we suggest using new-style caches from cachem. (#112)
 
 * Name clashes between function arguments and variables defined when memoising
   no longer occur (@egnha, #43).
@@ -7,10 +9,10 @@
 
 * Add `compress` option for non-memory caches (@coolbutuseless, #71).
 
-* Use absolute path in cache file system backend, so user can change working 
+* Use absolute path in cache file system backend, so user can change working
   directory after using relative path (@xhdong-umd, #51, #65)
 
-* Add `drop_cache()` to drop the cached result for particular arguments 
+* Add `drop_cache()` to drop the cached result for particular arguments
   (@richardkunze, #78)
 
 * Suppress messages of `aws.s3::head_object` within `cache_s3`'s `cache_has_key`